requirements.txt
annotated-types==0.7.0
anyio==4.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
openai==1.23.2
pdfminer.six==20221105
pdfplumber==0.10.2
pillow==10.4.0
pycparser==2.22
pydantic==2.10.6
pydantic-core==2.27.2
pypdfium2==4.30.1
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.13.2
